jobSearcherAgent/requirements.txt
crewai
langchain-openai
tavily-python
python-dotenv
pydantic
beautifulsoup4
selenium
webdriver-manager
requests
